backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
bcrypt==4.1.1
PyJWT==2.8.0
python-dotenv==1.0.0
qrcode[pil]==7.4.2
Werkzeug==3.0.1
Pillow==10.1.0
